Fix various VisualisationBuilder related conversion warnings

+++ source/app/ui/visualisations/visualisationchannel.cpp
@@ -29,7 +29,7 @@ void VisualisationChannel::reset()
 void VisualisationChannel::addValue(const QString& value)
 {
     _values.emplace_back(value);
-    _valueIndexMap[value] = static_cast<int>(_values.size()) - 1;
+    _valueIndexMap[value] = _values.size() - 1;
 }
 
 int VisualisationChannel::indexOf(const QString& value) const
